old-tests/submission/Microsoft/video/video_005.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: 'loop' is a boolean attribute.</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#attr-media-loop" />
  <meta name="assert" content="'loop' is a boolean attribute." />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
  <script type="text/javascript" src="../common/media.js"></script>
  
</head>
<body>
<pre>
'loop' is a boolean attribute.
</pre>
<pre>

http://www.w3.org/TR/html5/video.html#attr-media-loop

The loop attribute is a boolean attribute that, if specified, indicates 
that the media element is to seek back to the start of the media 
resource upon reaching the end. 

</pre>
<div id='log'></div>

<video id="video0">Your browser does not support media elements.</video>
<script type="text/javascript">    
  test(function ()
        {
     assert_equals(typeof document.getElementById("video0").loop, "boolean");
        }, "'loop' is a boolean attribute.");
    </script>

</body>
</html>
